01155543 - CTN Weekly Meeting: Thursday, 26-Mar-2020
We received some files containing additional data values.  This was received on the FTP server pointing to our UAT environment.

However, the XMLs for shipment does not have the new fields like PPS number, container, etc. Is this because the mapping changes on Axway side was not deployed in that environment.

Please see the file names that were sent from Cardinal to Axway.
If it was not processed with the new mappings, can you please reprocess and send the shipment XMLs?

git-svn-id: https://svn1.ecd.axway.int/svn/customers/US/BioMarin/TEST/CTN@1698 9ac7d299-a7d5-4a04-a3f5-b4a975ba8306